Update script.js
Display Question is mainly for display question  with multiple choice of Answers formate .

Every question has index and It will be displayed as per order of total questions.

Select Option functionality consider option button and option as  parameters.Evaluate the options if theycorrect or wrong by showing green and red color of option button which we selected.